+++ test/outline/test_outline_serializer.py
@@ -30,7 +30,7 @@ endobj
 7 0 obj
 <<
 /Count 1
-/Dest [3 0 R /XYZ 0 0.00 null]
+/Dest [3 0 R /XYZ 0 0 null]
 /First 8 0 R
 /Last 8 0 R
 /Next 9 0 R
@@ -41,7 +41,7 @@ endobj
 8 0 obj
 <<
 /Count 0
-/Dest [5 0 R /XYZ 0 0.00 null]
+/Dest [5 0 R /XYZ 0 0 null]
 /Parent 7 0 R
 /Title (Subtitle 1.1)
 >>
@@ -49,7 +49,7 @@ endobj
 9 0 obj
 <<
 /Count 2
-/Dest [7 0 R /XYZ 0 0.00 null]
+/Dest [7 0 R /XYZ 0 0 null]
 /First 10 0 R
 /Last 11 0 R
 /Parent 6 0 R
@@ -60,7 +60,7 @@ endobj
 10 0 obj
 <<
 /Count 0
-/Dest [9 0 R /XYZ 0 0.00 null]
+/Dest [9 0 R /XYZ 0 0 null]
 /Next 11 0 R
 /Parent 9 0 R
 /Title (Subtitle 2.1)
@@ -69,7 +69,7 @@ endobj
 11 0 obj
 <<
 /Count 0
-/Dest [11 0 R /XYZ 0 0.00 null]
+/Dest [11 0 R /XYZ 0 0 null]
 /Parent 9 0 R
 /Prev 10 0 R
 /Title (Subtitle 2.2)